+++ kliff/ml/training_wheels.py
@@ -1,7 +1,6 @@
 import numpy as np
 import torch
 from .graphs import KIMTorchGraphGenerator, KIMTorchGraph
-from torch_scatter import scatter
 from .descriptors import Descriptor
 import os
 
@@ -11,172 +10,10 @@ try:
 except ImportError:
     num_diff_available = False
 
+from .grad import *
+
 # TODO: Figure out way to incorporate original descriptor class
 # you will need to take manual jacobian vector product
-
-class TorchDescriptorFunction(torch.autograd.Function):
-    """
-    This class is a wrapper for the Descriptor class in kliff_torch.
-    It is used to compute the descriptor values and derivatives.
-    """
-    @staticmethod
-    def forward(ctx, DescFunCtx, configuration, coordinates):
-        """
-        Coordinates tensor needs to be passed separately .
-        Args:
-        :param ctx:
-        :param DescFunCtx:
-        :param coordinates: Coordinate tensor to accumulate gradients
-
-        Returns:
-
-        """
-        ctx.DescFunCtx = DescFunCtx
-        ctx.configuration = configuration
-        descriptor_tensor = DescFunCtx.forward(configuration)
-        descriptor_tensor = torch.from_numpy(descriptor_tensor)
-        descriptor_tensor.requires_grad_(True)
-        return descriptor_tensor
-
-    @staticmethod
-    def backward(ctx, grad_outputs):
-        """
-        Args:
-        :param ctx:
-        :param grad_outputs:
-        :return:
-        """
-        DescFunCtx = ctx.DescFunCtx
-        configuration = ctx.configuration
-        dE_dzeta = grad_outputs.numpy()
-        dE_dr = DescFunCtx.backward(configuration, dE_dzeta)
-        dE_dr = torch.from_numpy(dE_dr)
-        dE_dr.requires_grad_(True)
-        return None, None, dE_dr
-
-
-class TorchLegacyDescriptorFunction(torch.autograd.Function):
-    """
-    This class is a wrapper for the Descriptor class in kliff legacy descriptors.
-    """
-    @staticmethod
-    def forward(ctx, DescFunCtx, configuration, coordinates):
-        """
-        Coordinates tensor needs to be passed separately .
-        Args:
-        :param ctx:
-        :param DescFunCtx:
-        :param coordinates: Coordinate tensor to accumulate gradients
-
-        Returns:
-
-        """
-        ctx.DescFunCtx = DescFunCtx
-        ctx.configuration = configuration
-        descriptor_tensor, DescJacobian, _ = DescFunCtx.transform(configuration, fit_forces=True)
-        ctx.DescJacobian = DescJacobian
-        descriptor_tensor = torch.from_numpy(descriptor_tensor)
-        descriptor_tensor.requires_grad_(True)
-        return descriptor_tensor
-
-    @staticmethod
-    def backward(ctx, grad_outputs):
-        """
-        Args:
-        :param ctx:
-        :param grad_outputs:
-        :return:
-        """
-        DescFunCtx = ctx.DescFunCtx
-        configuration = ctx.configuration
-        dE_dzeta = grad_outputs.numpy()
-        dE_dr = np.einsum('ijk,ij->i', ctx.DescJacobian, dE_dzeta)
-        dE_dr = torch.from_numpy(dE_dr.reshape(-1, 3))
-        dE_dr.requires_grad_(True)
-        return None, None, dE_dr
-
-if num_diff_available:
-    # KIMModel wrappers, there will two kind of wrappers, one which computes the dataset,
-    # and the others which accept parameters and configurations as input.
-    class TorchKIMModelWrapper(torch.autograd.Function):
-        """
-        This class wraps the KIMModel for compatibility with torch autograd.
-        It uses numdifftools for differentiation over model parameters. This method would
-        not accept dataset as a parameter
-        """
-        @staticmethod
-        def forward(ctx, model, configuration, parameters:torch.Tensor):
-            """
-            Coordinates tensor needs to be passed separately .
-            :param ctx:
-            :param model:
-            :param configuration:
-            :param parameters:
-            :return:
-            """
-            model_param_fn = model.get_parameter_function(configuration)
-            ctx.model_param_fn = model_param_fn
-            ctx.parameters = parameters
-            ctx.grad_fn = nd.Gradient(model_param_fn)
-            return torch.tensor(model_param_fn(parameters.numpy()), requires_grad=True)
-
-        @staticmethod
-        def backward(ctx, grad_outputs):
-            """
-            Args:
-            :param ctx:
-            :param grad_outputs:
-            :return:
-            """
-            grad_fn = ctx.grad_fn
-            parameters = ctx.parameters
-            dE_dparam = grad_fn(parameters.numpy())
-            if type(dE_dparam) == np.float64 or dE_dparam.shape == ():
-                dE_dparam = np.array([dE_dparam])
-            input_grad = grad_outputs.numpy()
-            dE_dparam = np.dot(dE_dparam, input_grad)
-            dE_dparam = torch.from_numpy(dE_dparam)
-            dE_dparam.requires_grad_(True)
-            return None, None, dE_dparam
-
-
-class TorchGraphFunction(torch.autograd.Function):
-    """
-    This class is a wrapper for the KIMTorchGraph class in kliff_torch.
-    """
-    @staticmethod
-    def forward(ctx, GraphCtx:KIMTorchGraphGenerator, configuration, coordinate:torch.Tensor):
-        """
-        Coordinates tensor needs to be passed separately .
-        :param ctx:
-        :param GraphCtx:
-        :param configuration:
-        :param coordinate:
-        :return:
-        """
-        ctx.GraphCtx = GraphCtx
-        graph:KIMTorchGraph = GraphCtx.generate_graph(configuration)
-        outputs = [graph.species, graph.coords]
-        for i in range(graph.n_layers):
-            outputs.append(graph.__getattr__(f"edge_index{i}"))
-        outputs.append(graph.contributions)
-        ctx.graph = graph
-        return tuple(outputs)
-
-    @staticmethod
-    def backward(ctx, *grad_outputs):
-        """
-        Args:
-        :param ctx:
-        :param grad_outputs:
-        :return:
-        """
-        graph = ctx.graph
-        images = graph.images
-        d_coords = grad_outputs[1]
-        d_coords = scatter(d_coords, images, 0)
-        return None, None, d_coords
-
 
 class TrainingWheels(torch.nn.Module):
     """
@@ -186,13 +23,13 @@ class TrainingWheels(torch.nn.Module):
     @classmethod
     def init_graph(cls, model, cutoff, n_layers, species, model_returns_forces=False):
         """
-        Initialize a graph generator context
-        :param model:
-        :param cutoff:
-        :param n_layers:
-        :param species:
-        :param model_returns_forces:
-        :return:
+        Initialize a graph generator context and a TrainingWheels object.
+        :param model: TorchScript model to be trained
+        :param cutoff: Cutoff radius
+        :param n_layers: Number of convolution layers
+        :param species: List of atomic species
+        :param model_returns_forces: Whether the model is a gradient domain model
+        :return: :class:`kliff.ml.training_wheels.TrainingWheels`
         """
         kgg = KIMTorchGraphGenerator(species, cutoff, n_layers,as_torch_geometric_data=True)
         preprocessor = TorchGraphFunction()
@@ -203,15 +40,18 @@ class TrainingWheels(torch.nn.Module):
     def init_descriptor(cls, model, cutoff, descriptor_kind, species, hyperparams, model_returns_forces=False,
                         use_libdescriptor=True, **kwargs):
         """
-        Initialize a descriptor context
-        :param model:
-        :param cutoff:
-        :param descriptor_kind:
-        :param species:
-        :param model_returns_forces:
-        :param use_libdescriptor:
-        :param kwargs:
-        :return:
+        Initialize a TrainingWheels object along with a descriptor context. The generated descriptor context will be
+        passed to the forward function. The descriptor context is initialized with the given hyperparameters if
+        `use_libdescriptor` is True, otherwise the hyperparameters are passed to the descriptor context as kwargs.
+        This for flexibility of using different descriptor implementations.
+        :param model: TorchScript model to be trained
+        :param cutoff: Cutoff radius
+        :param descriptor_kind: Descriptor kind name obtained from :func:`kliff.ml.descriptor.show_available_descriptors()`
+        :param species: List of atomic species
+        :param model_returns_forces: Whether the model is a gradient domain model
+        :param use_libdescriptor: Whether to use the descriptor implementation in libdescriptor
+        :param kwargs: Hyperparameters for custom descriptor context
+        :return: :class:`kliff.ml.training_wheels.TrainingWheels` instance
         """
         if use_libdescriptor:
             descriptor_ctx = Descriptor(cutoff, species, descriptor_kind, hyperparams)
@@ -224,12 +64,12 @@ class TrainingWheels(torch.nn.Module):
     @classmethod
     def from_descriptor_instance(cls, model, descriptor, model_returns_forces=False):
         """
-        Initialize a descriptor context
-        :param model:
-        :param descriptor:
-        :param model_returns_forces:
-        :param use_libdescriptor:
-        :return:
+        Initialize a descriptor context from  an initialized descriptor context . The generated descriptor
+        context will be passed to the forward function.
+        :param model: TorchScript model to be trained
+        :param descriptor: :class:`kliff.descriptors.Descriptor` instance
+        :param model_returns_forces: Whether the model is a gradient domain model
+        :return: :class:`kliff.ml.training_wheels.TrainingWheels` instance
         """
         descriptor_ctx = descriptor
         preprocessor = TorchDescriptorFunction()
@@ -249,9 +89,11 @@ class TrainingWheels(torch.nn.Module):
 
     def forward(self, configuration):
         """
-
-        :param configuration:
-        :return:
+        Forward pass of the model. This will take in a :class:`kliff.dataset.Configuration` object, and return a
+        dictionary containing the energy, forces, and stress. The forces are negative of the gradient returned by the
+        model. As it returns forces as well, no backpropagation is needed.
+        :param configuration: :class:`kliff.dataset.Configuration` object or :class:`ase.Atoms` object
+        :return: Dictionary containing energy, forces, and stress
         """
         # coordinate_tensor = torch.from_numpy(self.descriptor_ctx.get_padded_coordinates(configuration))
         coordinate_tensor = torch.from_numpy(configuration.coords)
@@ -273,6 +115,17 @@ class TrainingWheels(torch.nn.Module):
         return self.parameters
 
     def save_kim_model(self, model_name):
+        """
+        Save the TorchScript model as the KIM ported model. It will also save the descriptor and graph
+        information if applicable. The model will be saved in the model_name directory. The model will be saved as
+        model.pt, and the descriptor and graph information will be saved as kim_model.param and kim_descriptor.dat
+        respectively. The saved model can be installed as
+        .. code-block:: bash
+            kim-api-collections-management install user/system  model_name
+
+        :param model_name: Model name (Recommended format: MD_000000000000_000)
+        :return:
+        """
         try:
             os.mkdir(model_name)
         except FileExistsError:
@@ -293,6 +146,12 @@ class TrainingWheels(torch.nn.Module):
 
     @staticmethod
     def _write_cmake_file(model_name, file_names):
+        """
+        Write a cmake file for the TorchMLModelDriver
+        :param model_name: string containing the name of the model
+        :param file_names: name of the files to to be included in model
+        :return:
+        """
         with open(f"{model_name}/CMakeLists.txt", "w") as f:
             f.write("cmake_minimum_required(VERSION 3.10)\n\n")
             f.write("list(APPEND CMAKE_PREFIX_PATH $ENV{KIM_API_CMAKE_PREFIX_DIR})\n")